Unify settings into one shared, daemon-aware dialog; delete monolith
One SettingsDialog (packages/ui/components/settings/SettingsDialog.tsx) now
serves every surface, adapting via a `daemonAvailable` flag instead of being
wired to the frontend app-store:

- Daemon/session present (frontend): full dialog, server-synced, all tabs.
- No daemon (standalone plan editor / portal): same dialog, cookie-only,
  daemon-only tabs hidden (AI, Hooks, git-name, legacyTabMode).

- New shared dialog + Dialog/Tabs primitives in packages/ui (radix-tabs dep).
- AppSettingsDialog is now a thin frontend adapter feeding the shared dialog
  its session context (mode/origin/apiBase) from the app-store.
- Plan editor's standalone built-in (AppHeader) renders the shared dialog
  (sessionContext=null, daemonAvailable=false); embedded path unchanged.
- configStore skips the server POST when no daemon/session base resolves
  (no doomed request in the portal); cookie writes unchanged.
- Deleted packages/ui/components/Settings.tsx (the last monolith).

Verified: packages/ui + apps/frontend typecheck clean; build:hook (verified
single-file) and build:portal both green; two adversarial reviews found nothing.
